// src/common/model/sso/StoreInfo.ts
export interface StoreInfo {
  store_name: string;
  store_short_name: string;
  store_code: string;
  postal_code: string;
  prefecture: string;
  address: string;
  search_index: string;
  phone_number: string;
  fax_number?: string;
  longitude: string;
  latitude: string;
  transportation?: string;
  news_from_store?: string;
  regular_holiday?: string;
  exist_parking: number;
  parking_comment?: string;
  can_credit_card: number;
  can_use_credit_card_brands?: string;
  can_electronic_money: number;
  can_electronic_money_brands?: string;
  can_bar_code_payment: number;
  can_bar_code_payment_brands?: string;
  monday_business_from_time: string;
  monday_business_end_time: string;
  tuesday_business_from_time: string;
  tuesday_business_end_time: string;
  wednesday_business_from_time: string;
  wednesday_business_end_time: string;
  thursday_business_from_time: string;
  thursday_business_end_time: string;
  friday_business_from_time: string;
  friday_business_end_time: string;
  saturday_business_from_time: string;
  saturday_business_end_time: string;
  sunday_business_from_time: string;
  sunday_business_end_time: string;
  national_holiday_business_from_time: string;
  national_holiday_business_end_time: string;
  business_time_display?: string;
  is_include_display_national_holiday: number;
  business_time_comment?: string;
  monday_reception_from_time: string;
  monday_reception_end_time: string;
  tuesday_reception_from_time: string;
  tuesday_reception_end_time: string;
  wednesday_reception_from_time: string;
  wednesday_reception_end_time: string;
  thursday_reception_from_time: string;
  thursday_reception_end_time: string;
  friday_reception_from_time: string;
  friday_reception_end_time: string;
  saturday_reception_from_time: string;
  saturday_reception_end_time: string;
  sunday_reception_from_time: string;
  sunday_reception_end_time: string;
  national_holiday_reception_from_time: string;
  national_holiday_reception_end_time: string;
  can_issue_terminal_certificate: number;
  use_terminal_certificate: number;
  can_edit_self_info: number;
  can_manage_staff: number;
  is_public: number;
  is_business_setting: number;
  exist_logo_image: number;
  exist_store_image: number;
  shop_id?: string;
  shop_pass?: string;
  site_id?: string;
  site_pass?: string;
  logo_image: string;
}